Introduce vmlinux.h header infrastructure
Anything but trivial BPF programs require the vmlinux.h header to know
existence and layout of miscellaneous Linux kernel types. Getting access
to this header file is painful enough for C code (requires availability
of bpftool and "shelling out" to it or an out-of-band mechanism for
putting it in place) and an outright nightmare for Rust code (basically
everything is a library here and expected to be self-contained [within
reason]; just relying on <stuff> to be available somewhere is a recipe
for disaster).
To that end, libbpf-bootstrap -- mostly a C code base -- vendors copies
of vmlinux.h for various architectures. libbpf-rs, a Rust code base, has
eventually introduced a test-only crate [0] that provides a usable Rust
interface and is trivially consumable from test code within the
repository. It is based on the libbpf-rs crate and its headers.

With upcoming changes another Rust code base -- blazesym [1] -- will
require vmlinux.h to be available for its testing infrastructure. To not
introduce a third copy of these header files, with this change we take a
first step towards unifying all of them under a single umbrella
repository. Specifically, we copy libbpf-rs's vmlinux workspace member
and expose it here. Subsequently, libbpf-rs will be able to depend on
this repository as opposed to having to manage these headers itself.
Similarly, for libbpf-bootstrap we can, if we so choose, switch over to
consuming this repository as a submodule, and including headers from it.

Note that, conceptually, there is no one vmlinux.h -- the file depends
on both the Linux kernel version as well as the configuration in use. As
such, there is every reason to believe that your workflow needs a
specific header. Existence of this repository does not go counter this
fact. Rather, we explicitly ship (at this point in time) a single
vmlinux.h for a specific version of the Linux kernel. The assumption
being that it will be good enough for various testing purposes, but it
may not suffice for each and every use case conceivable. We do *not*
recommend relying on this header file for anything but testing or other
development related tasks.
For that reason, we do not intend to publish the Rust crate. It will
need to be consumed as a git dependency, which has the nice effect that
having a non-dev git dependency prevents your crate from being published
to crates.io. It also forces users to specify a SHA-1 at which to
consume the contents. This is similar for consumption via git submodule,
which by design are "pinned".
Because of that, we have some leeway with respect to what kind of
changes we can apply here -- there is generally no guarantee of
backwards compatibility. Users consume header files at a given commit
and have to retest on upgrade/downgrade. That being said, none of this
precludes establishing additional header versioning in the future or
expanding the underlying kernel configuration.

[0] https://github.com/libbpf/libbpf-rs/commit/51fd1c6d70bd0bcb2eb358cf46fe60537ff3fcf9
[1] https://github.com/libbpf/blazesym

Signed-off-by: Daniel Müller <deso@posteo.net>